feat: :fire: Two files were created, one is called index.js and the other package.json and also the folder node_modules
the .json file defines the configuration of a GraphQL server project, specifying its name, version, and dependencies such as apollo-server, graphql, and uuid. In addition, it sets the project's entry point (index.js) and dependency settings required for operation.  The index.js file defines a schema with Person and Address data types, and provides resolvers for queries (personCount, allPersons, findPerson) and mutations (addPerson). The addPerson mutation allows you to add a person to the list if their name is unique. If the name already exists, it throws an input error (UserInputError). The server listens to requests on the generated URL and returns information about people and addresses.